// src/api/models/AppointmentSerializerRead.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AppointmentSerializerRead = {
    appointment_id?: number;
    tag_id?: number;
    date_start?: string;
    date_end?: string;
    note?: string;
    reminder?: string;
    property_id?: number;
    user_id_1?: number;
    user_id_2?: number;
};